# web/src/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/canonical-entities": {
      "get": {
        "operationId": "CanonicalEntitiesController_list",
        "parameters": [
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated fields to include; `_id` is always returned.",
            "schema": {
              "example": "name,status",
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of documents to skip (offset).",
            "schema": {
              "minimum": 0,
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of documents to return.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Field to sort by. Defaults to canonicalNameNormalized.",
            "schema": {
              "example": "canonicalNameNormalized",
              "type": "string",
              "enum": [
                "canonicalNameNormalized",
                "createdAt"
              ]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "description": "Sort direction. Defaults to asc.",
            "schema": {
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The tenant's registered canonical entities.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanonicalEntityListResponseDto"
                },
                "examples": {
                  "example": {
                    "summary": "One registered entity",
                    "value": {
                      "docs": [
                        {
                          "id": "65f1c2e4a1b2c3d4e5f6a7b8",
                          "canonicalName": "Northgate Business Park",
                          "aliases": [
                            "Northgate Bus. Park"
                          ],
                          "harvestedAliases": [
                            {
                              "alias": "Property",
                              "status": "proposed",
                              "quote": "Northgate Business Park (the \"Property\")",
                              "locator": {
                                "kind": "pdf-page",
                                "page": 4,
                                "extractorVersion": "pdf-1"
                              },
                              "documentVersionId": "65f1c2e4a1b2c3d4e5f6a7c9",
                              "harvestedAt": "2026-07-02T00:00:00.000Z"
                            }
                          ],
                          "createdAt": "2026-07-01T00:00:00.000Z"
                        }
                      ],
                      "count": 1
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "canonical-entities"
        ]
      },
      "post": {
        "operationId": "CanonicalEntitiesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCanonicalEntityRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The canonical entity row that was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanonicalEntityResponseDto"
                },
                "examples": {
                  "example": {
                    "summary": "Created entity",
                    "value": {
                      "id": "65f1c2e4a1b2c3d4e5f6a7b8",
                      "canonicalName": "Northgate Business Park",
                      "aliases": [
                        "Northgate Bus. Park"
                      ],
                      "harvestedAliases": [
                        {
                          "alias": "Property",
                          "status": "proposed",
                          "quote": "Northgate Business Park (the \"Property\")",
                          "locator": {
                            "kind": "pdf-page",
                            "page": 4,
                            "extractorVersion": "pdf-1"
                          },
                          "documentVersionId": "65f1c2e4a1b2c3d4e5f6a7c9",
                          "harvestedAt": "2026-07-02T00:00:00.000Z"
                        }
                      ],
                      "createdAt": "2026-07-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Caller does not hold the admin role required to author this registry.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Insufficient role",
                    "value": {
                      "statusCode": 403,
                      "message": "Insufficient role for this action",
                      "error": "Forbidden"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "An alias group with this name already exists for this tenant.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Duplicate canonical name",
                    "value": {
                      "statusCode": 409,
                      "message": "An alias group named 'Northgate Business Park' already exists for this tenant",
                      "error": "Conflict"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "canonical-entities"
        ]
      }
    },
...
    "/measures/{id}/confirm": {
      "post": {
        "operationId": "MeasuresController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasureEditsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The measure row after this confirmation. Triggers a synchronous rescan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasureResponseDto"
                },
                "examples": {
                  "example": {
                    "summary": "Confirmed measure",
                    "value": {
                      "id": "65f1c2e4a1b2c3d4e5f6a7b8",
                      "slug": "cap_rate",
                      "label": "Cap Rate",
                      "aliases": [
                        "Cap Rate",
                        "capitalization rate"
                      ],
                      "valueType": "percentage",
                      "canonicalUnit": "ratio",
                      "units": [
                        {
                          "id": "ratio",
                          "toCanonicalFactor": 1
                        },
                        {
                          "id": "percent",
                          "toCanonicalFactor": 0.01
                        }
                      ],
                      "toleranceKind": "absolute",
                      "tolerance": 0.0025,
                      "stalenessWindowMs": 15552000000,
                      "status": "confirmed",
                      "origin": "seed",
                      "proposedFrom": [],
                      "version": 2,
                      "createdAt": "2026-07-01T00:00:00.000Z",
                      "confirmedBy": "65f1c2e4a1b2c3d4e5f6a7c1",
                      "confirmedAt": "2026-07-02T00:00:00.000Z",
                      "lastRescan": {
                        "at": "2026-07-02T00:00:00.000Z",
                        "status": "completed",
                        "durationMs": 42,
                        "conflictsCreated": 0,
                        "skippedFactCount": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The merged definition (current row plus edits) fails validation.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "No factor-1 unit",
                    "value": {
                      "statusCode": 400,
                      "message": "exactly one unit must have toCanonicalFactor 1 and its id must equal canonicalUnit 'ratio'",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Caller does not hold the admin role required to author this registry.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Insufficient role",
                    "value": {
                      "statusCode": 403,
                      "message": "Insufficient role for this action",
                      "error": "Forbidden"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No measure with this id exists for the caller’s tenant.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Unknown id",
                    "value": {
                      "statusCode": 404,
                      "message": "Measure '65f1c2e4a1b2c3d4e5f6a7b8' not found",
                      "error": "Not Found"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "This measure is not in the proposed state and cannot be confirmed or rejected.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Already confirmed",
                    "value": {
                      "statusCode": 409,
                      "message": "Measure '65f1c2e4a1b2c3d4e5f6a7b8' is 'confirmed', not 'proposed'",
                      "error": "Conflict"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "measures"
        ]
      }
    },
...
    "/measures/{id}": {
      "patch": {
        "operationId": "MeasuresController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasureEditsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The measure row after this edit. Triggers a synchronous rescan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasureResponseDto"
                },
                "examples": {
                  "example": {
                    "summary": "Updated measure",
                    "value": {
                      "id": "65f1c2e4a1b2c3d4e5f6a7b8",
                      "slug": "cap_rate",
                      "label": "Cap Rate",
                      "aliases": [
                        "Cap Rate",
                        "capitalization rate"
                      ],
                      "valueType": "percentage",
                      "canonicalUnit": "ratio",
                      "units": [
                        {
                          "id": "ratio",
                          "toCanonicalFactor": 1
                        },
                        {
                          "id": "percent",
                          "toCanonicalFactor": 0.01
                        }
                      ],
                      "toleranceKind": "absolute",
                      "tolerance": 0.0025,
                      "stalenessWindowMs": 15552000000,
                      "status": "confirmed",
                      "origin": "seed",
                      "proposedFrom": [],
                      "version": 2,
                      "createdAt": "2026-07-01T00:00:00.000Z",
                      "confirmedBy": "65f1c2e4a1b2c3d4e5f6a7c1",
                      "confirmedAt": "2026-07-02T00:00:00.000Z",
                      "lastRescan": {
                        "at": "2026-07-02T00:00:00.000Z",
                        "status": "completed",
                        "durationMs": 42,
                        "conflictsCreated": 0,
                        "skippedFactCount": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The merged definition (current row plus edits) fails validation.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "No factor-1 unit",
                    "value": {
                      "statusCode": 400,
                      "message": "exactly one unit must have toCanonicalFactor 1 and its id must equal canonicalUnit 'ratio'",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Caller does not hold the admin role required to author this registry.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Insufficient role",
                    "value": {
                      "statusCode": 403,
                      "message": "Insufficient role for this action",
                      "error": "Forbidden"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No measure with this id exists for the caller’s tenant.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Unknown id",
                    "value": {
                      "statusCode": 404,
                      "message": "Measure '65f1c2e4a1b2c3d4e5f6a7b8' not found",
                      "error": "Not Found"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "This measure is not in the confirmed state and cannot be edited.",
            "content": {
              "application/json": {
                "examples": {
                  "example": {
                    "summary": "Still proposed",
                    "value": {
                      "statusCode": 409,
                      "message": "Measure '65f1c2e4a1b2c3d4e5f6a7b8' is 'proposed', not 'confirmed'",
                      "error": "Conflict"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "measures"
        ]
      }
    },
...
  },
  "info": {
    "title": "Evidence Ops API",
    "description": "Evidence Ops API specification",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "HarvestedAliasResponseDto": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "example": "Property",
            "description": "Alias as the document wrote it."
          },
          "status": {
            "type": "string",
            "example": "proposed",
            "enum": [
              "proposed",
              "applied",
              "revoked"
            ],
            "description": "Whether this alias resolves (applied), is recorded only (proposed), or has been rejected (revoked)."
          },
          "quote": {
            "type": "string",
            "example": "Northgate Business Park (the \"Property\")",
            "description": "Verbatim span of the document that defines this alias."
          },
          "locator": {
            "type": "object",
            "example": {
              "kind": "pdf-page",
              "page": 4,
              "extractorVersion": "pdf-1"
            },
            "description": "Where the quote sits in the document version it was read from."
          },
          "documentVersionId": {
            "type": "string",
            "example": "65f1c2e4a1b2c3d4e5f6a7b8",
            "description": "Document version the definition was read from."
          },
          "harvestedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01T00:00:00.000Z",
            "description": "When the definition was harvested."
          }
        },
        "required": [
          "alias",
          "status",
          "quote",
          "locator",
          "documentVersionId",
          "harvestedAt"
        ]
      },
      "CanonicalEntityResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65f1c2e4a1b2c3d4e5f6a7b8",
            "description": "Canonical entity identifier."
          },
          "canonicalName": {
            "type": "string",
            "example": "Northgate Business Park",
            "description": "Canonical display name this row registers alternate spellings against."
          },
          "aliases": {
            "example": [
              "Northgate Bus. Park"
            ],
            "description": "Alternate spellings that resolve to canonicalName.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "harvestedAliases": {
            "description": "Aliases read out of documents, each with the quote and locator defining it.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HarvestedAliasResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01T00:00:00.000Z",
            "description": "When this row was first authored."
          }
        },
        "required": [
          "id",
          "canonicalName",
          "aliases",
          "harvestedAliases",
          "createdAt"
        ]
      },
      "CanonicalEntityListResponseDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "example": 2,
            "description": "Total number of documents returned."
          },
          "docs": {
            "description": "The tenant's registered canonical entities.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CanonicalEntityResponseDto"
            }
          }
        },
        "required": [
          "count",
          "docs"
        ]
      },
...
      "CreateCanonicalEntityRequestDto": {
        "type": "object",
        "properties": {
          "canonicalName": {
            "type": "string",
            "example": "Northgate Business Park",
            "description": "Canonical display name this row registers alternate spellings against."
          },
          "aliases": {
            "example": [
              "Northgate Bus. Park"
            ],
            "description": "Alternate spellings that should resolve to canonicalName.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "canonicalName"
        ]
      },
...
      "MeasureResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65f1c2e4a1b2c3d4e5f6a7b8",
            "description": "Measure identifier."
          },
          "slug": {
            "type": "string",
            "example": "cap_rate",
            "description": "Slug the ExtractedFact.factKey.metric join key carries."
          },
          "label": {
            "type": "string",
            "example": "Cap Rate",
            "description": "Human-readable label."
          },
          "aliases": {
            "example": [
              "Cap Rate",
              "capitalization rate"
            ],
            "description": "Header/phrase forms matched case-insensitively.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "valueType": {
            "type": "string",
            "example": "percentage",
            "enum": [
              "currency",
              "percentage",
              "area",
              "duration",
              "count"
            ],
            "description": "The kind of value this measure holds."
          },
          "canonicalUnit": {
            "type": "string",
            "example": "ratio",
            "description": "The unit every value converts to."
          },
          "units": {
            "example": [
              {
                "id": "ratio",
                "toCanonicalFactor": 1
              },
              {
                "id": "percent",
                "toCanonicalFactor": 0.01
              }
            ],
            "description": "Convertible units. Exactly one carries toCanonicalFactor 1, matching canonicalUnit.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "toleranceKind": {
            "type": "string",
            "example": "absolute",
            "enum": [
              "absolute",
              "relative"
            ],
            "description": "How tolerance is applied when comparing values."
          },
          "tolerance": {
            "type": "number",
            "example": 0.0025,
            "description": "Disagreement threshold before two values conflict."
          },
          "authorityOrder": {
            "example": [
              "pm-export",
              "spreadsheet"
            ],
            "description": "Source classes ranked by authority, most authoritative first.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "stalenessWindowMs": {
            "type": "number",
            "example": 15552000000,
            "description": "How long an observed value stays current for this measure, in milliseconds."
          },
          "status": {
            "type": "string",
            "example": "confirmed",
            "enum": [
              "proposed",
              "confirmed",
              "rejected"
            ],
            "description": "Measure lifecycle status."
          },
          "origin": {
            "type": "string",
            "example": "seed",
            "enum": [
              "seed",
              "header",
              "manual"
            ],
            "description": "How this measure came to exist."
          },
          "proposedFrom": {
            "example": [],
            "description": "Header evidence recorded while this measure was 'proposed' — empty for a seed or manually authored row.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "version": {
            "type": "number",
            "example": 1,
            "description": "Bumped on every confirm/update. Seed rows are the only rows that ever read 1 while confirmed."
          },
          "confirmedBy": {
            "type": "string",
            "example": "65f1c2e4a1b2c3d4e5f6a7c1",
            "description": "User id who confirmed this measure."
          },
          "confirmedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01T00:00:00.000Z",
            "description": "When this measure was confirmed."
          },
          "rejectedBy": {
            "type": "string",
            "example": "65f1c2e4a1b2c3d4e5f6a7c1",
            "description": "User id who rejected this measure."
          },
          "rejectedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01T00:00:00.000Z",
            "description": "When this measure was rejected."
          },
          "rejectedReason": {
            "type": "string",
            "example": "duplicate",
            "description": "Why this measure was rejected."
          },
          "lastRescan": {
            "type": "object",
            "example": {
              "at": "2026-07-01T00:00:00.000Z",
              "status": "completed",
              "durationMs": 42,
              "conflictsCreated": 0
            },
            "description": "Outcome of the synchronous conflict rescan the last confirm/update triggered. Absent until this measure has been confirmed or edited at least once."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01T00:00:00.000Z",
            "description": "When this row was first created."
          }
        },
        "required": [
          "id",
          "slug",
          "label",
          "aliases",
          "valueType",
          "canonicalUnit",
          "units",
          "toleranceKind",
          "tolerance",
          "status",
          "origin",
          "proposedFrom",
          "version",
          "createdAt"
        ]
      },
...
      "MeasureUnitRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "sf",
            "description": "Unit identifier, matched case-sensitively against a fact’s reported unit."
          },
          "toCanonicalFactor": {
            "type": "number",
            "example": 1,
            "description": "Multiplicative factor converting one of this unit into the canonicalUnit."
          }
        },
        "required": [
          "id",
          "toCanonicalFactor"
        ]
      },
      "MeasureEditsRequestDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "Tenant Occupancy Share",
            "description": "Human-readable label."
          },
          "aliases": {
            "example": [
              "Occupancy %"
            ],
            "description": "Replaces the full alias list. Omit to leave the current aliases untouched.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "valueType": {
            "type": "string",
            "example": "percentage",
            "enum": [
              "currency",
              "percentage",
              "area",
              "duration",
              "count"
            ],
            "description": "The kind of value this measure holds."
          },
          "canonicalUnit": {
            "type": "string",
            "example": "ratio",
            "description": "The unit every value converts to. Must name one of units below."
          },
          "units": {
            "description": "Replaces the full unit list. Exactly one unit must have toCanonicalFactor 1 and its id must equal canonicalUnit.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeasureUnitRequestDto"
            }
          },
          "toleranceKind": {
            "type": "string",
            "example": "absolute",
            "enum": [
              "absolute",
              "relative"
            ],
            "description": "How tolerance is applied when comparing values."
          },
          "tolerance": {
            "type": "number",
            "example": 0.02,
            "description": "Disagreement threshold before two values conflict."
          },
          "authorityOrder": {
            "type": "array",
            "example": [
              "pm-export",
              "spreadsheet"
            ],
            "description": "Source classes ranked by authority, most authoritative first.",
            "items": {
              "type": "string",
              "enum": [
                "crm-export",
                "pm-export",
                "spreadsheet",
                "memo",
                "report"
              ]
            }
          },
          "stalenessWindowMs": {
            "type": "number",
            "example": 7776000000,
            "description": "How long an observed value stays current for this measure, in milliseconds."
          }
        }
      },
...
    }
  }
}